requirements.txt
requests
beautifulsoup4
pandas
google-api-python-client
google-auth-httplib2
google-auth-oauthlib